Move ColorPalette utils to separate file
