+++ packages/components/src/color-palette/test/utils.ts
@@ -4,7 +4,7 @@
 import {
 	extractColorNameFromCurrentValue,
 	showTransparentBackground,
-} from '../';
+} from '../utils';
 
 describe( 'ColorPalette: Utils', () => {
 	describe( 'extractColorNameFromCurrentValue', () => {
